chore/100/rename-core-to-engine: renamed core to engine for clarity
