+++ packages/library/src/lib/plugins/backend.ts
@@ -1,4 +1,4 @@
-import { DATASTAR_STR } from '../core'
+import { DATASTAR_STR } from '../engine'
 import { fetchEventSource, FetchEventSourceInit } from '../external/fetch-event-source'
 import { idiomorph } from '../external/idiomorph'
 import { Signal } from '../external/preact-core'
